// src/app/models/userWiseFile.model.ts
export class UserWiseFile{
    Caption:string;
    UploadedFileName:string;
    UniqueFileName:string;
    UniqueFileID:string;
    FileUploadDate:string;
    FilePath:string;    
    ThumbnailPath:string;
    Tag1: string;
    Tag2: string;
    Tag3: string;
    Tag4: string;
    Tag5: string;
    FolderID: string;
    IsSmartRxExists: boolean;
}